+++ cli/tests/test_acls.rs
@@ -49,6 +49,9 @@ fn test_diff() {
     Modified regular file z-last:
        1    1: foobar
     [EOF]
+    ------- stderr -------
+    Tracking added-secret
+    [EOF]
     ");
     let output = work_dir.run_jj(["diff", "--summary"]);
     insta::assert_snapshot!(output.normalize_backslash(), @r"
@@ -119,6 +122,9 @@ fn test_file_list_show() {
     a-first
     secret
     z-last
+    [EOF]
+    ------- stderr -------
+    Tracking a-first, secret, z-last
     [EOF]
     ");