+++ cli/tests/test_copy_detection.rs
@@ -31,5 +31,8 @@ fn test_simple_rename() {
         work_dir.run_jj(["debug", "copy-detection"]).normalize_backslash(), @r"
     original -> modified
     [EOF]
+    ------- stderr -------
+    Tracking modified
+    [EOF]
     ");
 }