+++ cli/tests/test_file_show_command.rs
@@ -31,6 +31,9 @@ fn test_show() {
     insta::assert_snapshot!(output, @r"
     a
     [EOF]
+    ------- stderr -------
+    Tracking dir/file2
+    [EOF]
     ");
 
     // Defaults to printing the working-copy version
@@ -139,6 +142,7 @@ fn test_show_symlink() {
     --- symlink1 [symlink]
     [EOF]
     ------- stderr -------
+    Tracking dir/file2, file1, symlink1
     Warning: Path 'symlink1' exists but is not a file
     [EOF]
     ");