+++ cli/tests/test_show_command.rs
@@ -63,6 +63,9 @@ fn test_show_basic() {
        2    3: baz quxquux
     Modified regular file file3 (file1 => file3):
     [EOF]
+    ------- stderr -------
+    Tracking file3
+    [EOF]
     ");
 
     let output = work_dir.run_jj(["show", "--context=0"]);
@@ -268,6 +271,9 @@ fn test_show_with_template_no_patch() {
     insta::assert_snapshot!(output, @r"
     a new commit
     [EOF]
+    ------- stderr -------
+    Tracking file1
+    [EOF]
     ");
 }
 
@@ -289,6 +295,9 @@ fn test_show_with_no_patch() {
 
         a new commit
 
+    [EOF]
+    ------- stderr -------
+    Tracking file1
     [EOF]
     ");
 }