+++ cli/tests/test_squash_command.rs
@@ -348,6 +348,7 @@ fn test_squash_partial() {
     let output = work_dir.run_jj(["squash", "-i", "file1", "file3"]);
     insta::assert_snapshot!(output, @r"
     ------- stderr -------
+    Tracking file3
     Rebased 1 descendant commits
     Working copy  (@) now at: mzvwutvl 3615d80e c | (no description set)
     Parent commit (@-)      : kkmpptxz 037106c4 b | (no description set)
@@ -1186,6 +1187,9 @@ fn test_squash_from_multiple_partial_no_op() {
     ├─╯
     ○  93d495c46d89 a
     ◆  000000000000 (empty)
+    [EOF]
+    ------- stderr -------
+    Tracking d
     [EOF]
     ");
     let setup_opid = work_dir.current_operation_id();